app.py
import streamlit as st
import cv2
import numpy as np
from PIL import Image
import os
import tempfile

# Configuración de la página
st.set_page_config(
    page_title="Sistema de Playa de Estacionamiento Automatizada",
    page_icon="🚗",
    layout="wide"
)

# Título principal
st.title("🚗 Sistema de Playa de Estacionamiento Automatizada")
st.markdown("---")

# Sección 1: Visualización de la Aplicación
with st.container():
    st.header("1. Visualización de la Aplicación")
    col1, col2 = st.columns(2)
    
    with col1:
        st.subheader("Interfaz Principal")
        st.image("https://via.placeholder.com/600x400/4B8BBE/FFFFFF?text=Interfaz+Principal", 
                caption="Interfaz principal del sistema de estacionamiento")
    
    with col2:
        st.subheader("Funcionamiento en Tiempo Real")
        st.video("https://www.youtube.com/watch?v=example")
        st.caption("Demostración del sistema en funcionamiento")

# Sección 2: Detección de Placas y Licencias
with st.container():
    st.header("2. Detección de Placas y Licencias")
    
    tab1, tab2 = st.tabs(["Detección de Placas", "Detección de Licencias"])
    
    with tab1:
        st.subheader("Proceso de Detección de Placas")
        st.image("https://via.placeholder.com/800x300/306998/FFFFFF?text=Proceso+de+Detección+de+Placas",
                caption="1. Captura de imagen → 2. Procesamiento → 3. Extracción de texto → 4. Validación")
    
    with tab2:
        st.subheader("Proceso de Detección de Licencias")
        st.image("https://via.placeholder.com/800x300/FFD43B/000000?text=Proceso+de+Detección+de+Licencias",
                caption="1. Captura de imagen → 2. Procesamiento OCR → 3. Extracción de datos → 4. Validación")

# Sección 3: Paneles de Control y Monitoreo
with st.container():
    st.header("3. Paneles de Control y Monitoreo")
    
    col1, col2 = st.columns(2)
    
    with col1:
        st.subheader("Dashboard en Tiem Real")
        st.image("https://via.placeholder.com/600x400/4B8BBE/FFFFFF?text=Dashboard+de+Monitoreo",
                caption="Panel de control con métricas en tiempo real")
    
    with col2:
        st.subheader("Estadísticas")
        st.metric("Vehículos en estacionamiento", "24")
        st.metric("Espacios disponibles", "76")
        st.metric("Ingresos del día", "$1,250.00")
        st.metric("Promedio de estancia", "45 min")

# Sección 4: Funcionalidades Adicionales
with st.container():
    st.header("4. Otras Funcionalidades")
    
    features = {
        "🔒 Control de Acceso": "Registro automatizado de entradas y salidas",
        "📊 Reportes": "Generación de reportes diarios, semanales y mensuales",
        "📱 Notificaciones": "Alertas por correo o SMS para recordatorios",
        "🔍 Historial": "Registro completo de todos los vehículos estacionados",
        "⚙️ Configuración": "Personalización de tarifas y horarios"
    }
    
    for feature, description in features.items():
        with st.expander(feature):
            st.write(description)

# Sección de demostración (opcional)
with st.sidebar:
    st.header("Prueba el Sistema")
    uploaded_file = st.file_uploader("Sube una imagen de placa o licencia:", 
                                   type=["jpg", "jpeg", "png"])
    
    if uploaded_file is not None:
        image = Image.open(uploaded_file)
        st.image(image, caption="Imagen cargada", use_column_width=True)
        
        if st.button("Procesar Imagen"):
            with st.spinner('Procesando...'):
                # Aquí iría el código de procesamiento real
                st.success("¡Procesamiento completado!")
                st.write("Texto detectado: ABC-123")

# Pie de página
st.markdown("---")
st.caption("© 2024 Sistema de Playa de Estacionamiento Automatizada - Todos los derechos reservados")
