requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
torch
torchvision
numpy
pytesseract>=0.3.10
matplotlib>=3.4.0
Pillow>=10.0.0
imutils>=0.5.4
easyocr>=1.7.0
